feat: imporve structure, add failure metric, and predefined questions
