+++ ariadne_ai/evaluators/rag/answer_relevance_evaluator.py
@@ -1,54 +1,44 @@
-from .summarization_evaluator import SummarizationEvaluator
-from ..loaders.summarization_loader import SummarizationLoader
-from ..metrics.aggreement_score import AgreementScore 
-from ..metrics.contradiction_score import ContradictionScore
-from ..metrics.hallucination_score import HallucinationScore 
-from ..publishers.publisher_log import PublisherLog
-from ..llms.question_generator import QuestionGenerator
-from ..llms.question_answerer import QuestionAnswerer
+from .rag_evaluator import RagEvaluator
+from ...loaders.rag_loader import RagLoader
+from ...metrics.rag.answer_relevance_failure import AnswerRelevanceFailure 
+from ...publishers.publisher_log import PublisherLog
+from ...llms.rag.answer_relevance import AnswerRelevance
 
-class SummarizationHallucinationEvaluator(SummarizationEvaluator):
+class AnswerRelevanceEvaluator(RagEvaluator):
     """
-    Evaluator for hallucinations in text summarizations. 
+    Evaluator for answer relevance in rag chatbot. 
 
     Attributes:
         dataset: Dataset containing instances for evaluation.
-        question_generator: Question generator based on summaries.
-        question_answerer: Question answerer for the questions based on documents/summaries.
+        context_relevance_evaluator: Evaluator for context relevance.
         publisher_log: JSON publisher to save the evaluation logs.
         performance_report_filename: txt file to save the perfrormance of a batch
         metrics: List of metrics to evaluate.
         logs: List to accumulate evaluation results for each instance.
-        n_questions: Number of questions to be generated for each summary.
     """
-
+    # Chamge metric
     metric_str_to_class = {
-        'agreement_score': AgreementScore,
-        'hallucination_score': HallucinationScore,
-        'contradiction_score': ContradictionScore
+        'answer_relevance_failure': AnswerRelevanceFailure
     }
 
-    def __init__(self, loader, log_filepath='data/logs/log_sum_hal_eval.json', log_format = 'json', performance_filepath = 'data/logs/perf_sum_hal_eval.txt',  n_questions=5, 
-                 llm_model='gpt-3.5-turbo', metrics=['agreement_score', 'hallucination_score', 'contradiction_score'], open_ai_key = None):
+    def __init__(self, loader, log_filepath='data/logs/log_rag_ans_rel_eval.json', log_format = 'json', performance_filepath = 'data/logs/perf_rag_ans_rel_eval.txt', 
+                 llm_model='gpt-3.5-turbo', metrics=['answer_relevance_failure'], open_ai_key = None):
         """
         Initialize the evaluator with given parameters.
 
         Args:
         - loader: An instance of SummarizationLoader.
         - log_filepath: Path to save the logs.
-        - n_questions: Number of questions to generate for summaries.
         - llm_model: Language model to be used.
         - metrics: List of metrics for evaluation.
         """
-        if not isinstance(loader, SummarizationLoader):
-            raise TypeError("Loader must be an instance of SummarizationLoader")
+        if not isinstance(loader, RagLoader):
+            raise TypeError("Loader must be an instance of RagLoader")
         # Load data
         self.dataset = loader.processed_dataset
         # Intialize LLMs
         self.llm_model = llm_model
-        self.n_questions = n_questions
-        self.question_generator = QuestionGenerator(llm_model, n_questions,  open_ai_key)
-        self.question_answerer = QuestionAnswerer(llm_model, open_ai_key)
+        self.answer_relevance_evaluator = AnswerRelevance(llm_model, open_ai_key)
         # Initialize logging
         self.log_format = log_format
         if(log_format is not None):
@@ -64,38 +54,31 @@ class SummarizationHallucinationEvaluator(SummarizationEvaluator):
 
     def _evaluate_element(self, instance):
         """Evaluate an instance for hallucination."""
-        document = instance['document']
-        summary = instance['summary']
+        question = instance['question']
+        answer = instance['answer']
         if 'label' in instance:
             label = instance['label']
         else:
             label = 'overall'
-        
-        # Generate questions based on summary
-        questions = self.question_generator.generate(summary)
-        
-        # Get answers from document and summary
-        answers_doc = self.question_answerer.answer(questions, document)
-        answers_sum = self.question_answerer.answer(questions, summary)
+        # Run LLM Evaluator for faithfullness
+        anw_rel_eval = self.answer_relevance_evaluator.evaluate(question, answer)
+      
         metric_results = {}
-        print(answers_doc, answers_sum, questions)
         # Compute metrics
-        if( answers_doc is None or answers_sum is None or questions is None):
+        if( anw_rel_eval is None):
             metric_results['evaluation'] = 'undefined'
         else:
-            for score in self.metrics:
-                metric_class = self.metric_str_to_class.get(score)
-                metric_result = metric_class.compute(answers_doc, answers_sum, self.n_questions)
-                metric_results[score] = metric_result
-                self.update_metric_aggr(score, label, metric_result)
+            for metric in self.metrics:
+                metric_class = self.metric_str_to_class.get(metric)
+                metric_result, explanation = metric_class.compute(anw_rel_eval)
+                metric_results[metric] = metric_result
+                metric_results['reason'] = explanation
+                self.update_metric_aggr(metric, label, metric_result)
             self.n_instances = self.n_instances +1
             self.label_counts[label] = self.label_counts.get(label, 0) + 1
         return {
-            'document': document,
-            'summary': summary,
-            'questions': questions,
-            'answers_doc': answers_doc,
-            'answers_sum': answers_sum,
+            'question': question,
+            'answer': answer,
             'label': label,
             **metric_results
         }
@@ -137,7 +120,6 @@ class SummarizationHallucinationEvaluator(SummarizationEvaluator):
         """Generate a performance report and save it to the provided filename."""
         avg_scores = self.compute_average_scores()
         with open(filename, 'w') as f:
-            f.write(f"Number of Questions: {self.n_questions}\n")
             f.write(f"Model: {self.llm_model}\n")
             f.write("\nNumber of instances per Label:\n")
             for label, cnt in self.label_counts.items():
@@ -148,8 +130,6 @@ class SummarizationHallucinationEvaluator(SummarizationEvaluator):
                 for label, avg in avg_scores[metric].items():
                     f.write(f"{label}: {avg}\n")
 
-
-
     def run(self):
         """Evaluate all instances in the dataset."""
         for instance in self.dataset:
@@ -159,9 +139,3 @@ class SummarizationHallucinationEvaluator(SummarizationEvaluator):
         if(self.log_format is not None):
             self.publisher_log.write(self.logs)
         return self.logs
-
-    
-
-
-
-
