+++ ariadne_ai/metrics/rag/context_relevance_failure.py
@@ -1,4 +1,4 @@
-from .metric import Metric
+from ..metric import Metric
 
 class ContextRelevanceFailure(Metric):
     """
