+++ ariadne_ai/metrics/rag/faithfullness_failure.py
@@ -1,4 +1,5 @@
-from .metric import Metric
+from ..metric import Metric
+
 
 class FaithfullnessFailure(Metric):
     """
