+++ ariadne_ai/metrics/text_summarization/aggreement_score.py
@@ -1,4 +1,4 @@
-from .metric import Metric
+from ..metric import Metric
 
 class AgreementScore(Metric):
     """
@@ -9,7 +9,7 @@ class AgreementScore(Metric):
     """
 
     @staticmethod
-    def _compute_metric(answers_src, answers_sum):
+    def _compute_metric(answers_src, answers_sum, questions):
         """
         Computes the number of matches between the answers from source and summary.
 
@@ -23,13 +23,16 @@ class AgreementScore(Metric):
         answers_src_ls = list(answers_src.values())
         answers_sum_ls = list(answers_sum.values())
         n_matches = 0
-        for ans_src, ans_sum in zip(answers_src_ls, answers_sum_ls):
+        aggr_questions = []
+        for idx, (ans_src, ans_sum) in enumerate(zip(answers_src_ls, answers_sum_ls)):
             if ans_src.strip().lower() == ans_sum.strip().lower():
                 n_matches += 1
-        return n_matches
+                aggr_question = questions[f"question {idx+1}"]
+                aggr_questions.append(f'{aggr_question}')
+        return n_matches,aggr_questions
 
     @staticmethod
-    def compute(answers_src, answers_sum, n_questions):
+    def compute(answers_src, answers_sum, questions, n_questions):
         """
         Computes the agreement score.
 
@@ -41,6 +44,11 @@ class AgreementScore(Metric):
         Returns:
             float: Agreement score.
         """
-        n_matches = AgreementScore._compute_metric(answers_src, answers_sum)
+        n_matches, aggr_questions = AgreementScore._compute_metric(answers_src, answers_sum, questions) 
+        explanation = aggr_questions
         aggr_score = n_matches / n_questions
-        return aggr_score
+        return aggr_score, explanation, None
+    
+
+
+
