+++ data/logs/log_sum_hal_eval_gpt_35_questions_5.json
@@ -24,9 +24,19 @@
             "question 5": "Unknown"
         },
         "label": NaN,
+        "hallucination_failure": 1,
+        "reason_hallucination_failure": [
+            "Was a gas leak found at the fracking site in Lancashire?",
+            "Is the fracking operation still ongoing in Lancashire?",
+            "Was the gas leak the reason for the suspension of the fracking operation?"
+        ],
+        "contradiction_failure": 0,
+        "reason_contradiction_failure": [],
         "agreement_score": 0.4,
-        "hallucination_score": 0.6,
-        "contradiction_score": 0.0
+        "reason_agreement_score": [
+            "Was a fracking operation in Lancashire suspended?",
+            "Has the gas leak been resolved at the fracking site in Lancashire?"
+        ]
     },
     {
         "document": "Tries from Macgraff Leuluai, Ah Van and Chris Bridge gave Widnes a 14-6 half-time lead, with Cas loanee Paddy Flynn responding against his parent club.\nFlynn scored another early in the second, but Ah Van, Lloyd White (2) and Joe Mellor tries made it 38-12.\nDespite three Cas tries in four minutes - Larne Patrick, Luke Gale and Flynn's hat-trick score - Widnes held on.\nThe Vikings' victory had looked assured when Mellor took advantage of Castleford indecision on a short kick-off, running straight through to help them to a 26-point lead.\nBut, having lost nine of their previous 10 matches, their defence creaked as they let Patrick, who has just signed a three-year deal to move from Huddersfield after his loan spell, in for a try.\nGale and Flynn - on loan at Castleford from the Vikings until the end of the season - made it a nervy finish, but the Tigers suffered their fourth consecutive home league defeat.\nWidnes moved above Wakefield into seventh, while Castleford remain in sixth, one point above the Vikings.\nCastleford: Solomona, Monaghan, Crooks, Webster, Flynn, McShane, Gale, Jewitt, Milner, Patrick, Moors, McMeeken, Massey.\nReplacements: Millington, Springer, Tickle, Cook.\nWidnes: Hanbury, Thompson, Runciman, Bridge, Ah Van, Mellor,Brown, Cahill, White, Buchanan, Whitley, Dean, Leuluai.\nReplacements: J. Chapelhow, Manuokafoa, Walker, Heremaia.\nReferee: J Smith",
@@ -53,25 +63,35 @@
             "question 5": "Unknown"
         },
         "label": "intrinsic",
+        "hallucination_failure": 0,
+        "reason_hallucination_failure": [],
+        "contradiction_failure": 1,
+        "reason_contradiction_failure": [
+            "Did Widnes suffer their first defeat of the season?",
+            "Did Widnes suffer any defeats before this game?"
+        ],
         "agreement_score": 0.6,
-        "hallucination_score": 0.0,
-        "contradiction_score": 0.4
+        "reason_agreement_score": [
+            "Did Widnes win at home?",
+            "Did Widnes win against Widnes?",
+            "Did Widnes win their first game of the season?"
+        ]
     },
     {
         "document": "More than 20 competitors are due to take part in Saturday's event in Carrbridge in the Cairngorms.\nCarvers from Scotland, England, Ireland as well as continental Europe, Canada and the US regularly enter the competition.\nThey can use only chainsaws - any other carving tool is banned - to carve logs into the shapes of animals and people.\nLast year's winner Pete Bowsher, from Moffat, is to return to defend the title, and retain the first place winner's prize the Claymore Trophy.",
         "summary": "this year\\'s northern ireland car championships will take place at this year\\'s ballymena festival.",
         "questions": {
-            "question 1": "Will this year's northern ireland car championships take place at this year's ballymena festival?",
+            "question 1": "Will the northern ireland car championships take place at this year's ballymena festival?",
             "question 2": "Is this year's northern ireland car championships happening in Northern Ireland?",
             "question 3": "Is the ballymena festival happening this year?",
-            "question 4": "Are the northern ireland car championships a part of the ballymena festival?",
-            "question 5": "Will the ballymena festival host any other events besides the northern ireland car championships?"
+            "question 4": "Will the car championships take place at a different location?",
+            "question 5": "Is the ballymena festival exclusively for car championships?"
         },
         "answers_doc": {
-            "question 1": "No",
+            "question 1": "Unknown",
             "question 2": "Unknown",
             "question 3": "Unknown",
-            "question 4": "No",
+            "question 4": "Unknown",
             "question 5": "Unknown"
         },
         "answers_sum": {
@@ -82,8 +102,18 @@
             "question 5": "Unknown"
         },
         "label": "extrinsic",
-        "agreement_score": 0.4,
-        "hallucination_score": 0.2,
-        "contradiction_score": 0.4
+        "hallucination_failure": 1,
+        "reason_hallucination_failure": [
+            "Will the northern ireland car championships take place at this year's ballymena festival?",
+            "Is this year's northern ireland car championships happening in Northern Ireland?"
+        ],
+        "contradiction_failure": 0,
+        "reason_contradiction_failure": [],
+        "agreement_score": 0.6,
+        "reason_agreement_score": [
+            "Is the ballymena festival happening this year?",
+            "Will the car championships take place at a different location?",
+            "Is the ballymena festival exclusively for car championships?"
+        ]
     }
 ]
