+++ data/logs/perf_sum_hal_eval_gpt_35_questions_2.txt
@@ -7,20 +7,20 @@ intrinsic: 1
 extrinsic: 1
 total: 3
 
+Average hallucination_failure per Label:
+nan: 1.0
+intrinsic: 0.0
+extrinsic: 0.0
+overall: 0.3333333333333333
+
+Average contradiction_failure per Label:
+nan: 0.0
+intrinsic: 1.0
+extrinsic: 1.0
+overall: 0.6666666666666666
+
 Average agreement_score per Label:
 nan: 0.5
 intrinsic: 0.5
 extrinsic: 0.0
 overall: 0.3333333333333333
-
-Average hallucination_score per Label:
-nan: 0.5
-intrinsic: 0.0
-extrinsic: 0.0
-overall: 0.16666666666666666
-
-Average contradiction_score per Label:
-nan: 0.0
-intrinsic: 0.5
-extrinsic: 1.0
-overall: 0.5
