+++ data/logs/perf_sum_hal_eval_gpt_35_questions_5.txt
@@ -7,20 +7,20 @@ intrinsic: 1
 extrinsic: 1
 total: 3
 
+Average hallucination_failure per Label:
+nan: 1.0
+intrinsic: 0.0
+extrinsic: 1.0
+overall: 0.6666666666666666
+
+Average contradiction_failure per Label:
+nan: 0.0
+intrinsic: 1.0
+extrinsic: 0.0
+overall: 0.3333333333333333
+
 Average agreement_score per Label:
 nan: 0.4
 intrinsic: 0.6
-extrinsic: 0.4
-overall: 0.4666666666666666
-
-Average hallucination_score per Label:
-nan: 0.6
-intrinsic: 0.0
-extrinsic: 0.2
-overall: 0.26666666666666666
-
-Average contradiction_score per Label:
-nan: 0.0
-intrinsic: 0.4
-extrinsic: 0.4
-overall: 0.26666666666666666
+extrinsic: 0.6
+overall: 0.5333333333333333
