+++ run_experiment_rag.py
@@ -1,7 +1,7 @@
-from ariadne_ai.loaders.rag_generation_loader import RagGenerationLoader
-from ariadne_ai.evaluators.rag_faithfulness_evaluator import RagFaithfulnessEvaluator
-from ariadne_ai.evaluators.rag_context_relevance_evaluator import RagContextRelevanceEvaluator
-from ariadne_ai.evaluators.rag_answer_relevance_evaluator import RagAnswerRelevanceEvaluator
+from ariadne_ai.loaders.rag_loader import RagLoader
+from ariadne_ai.evaluators.rag.faithfulness_evaluator import FaithfulnessEvaluator
+from ariadne_ai.evaluators.rag.context_relevance_evaluator import ContextRelevanceEvaluator
+from ariadne_ai.evaluators.rag.answer_relevance_evaluator import AnswerRelevanceEvaluator
 # Constants and Configurations
 
 # Path to the input file containing raw data.
@@ -12,41 +12,36 @@ PERF_FILEPATH = 'data/logs/perf_rag_faith_eval.txt'
 # OpenAI API key (should be kept confidential).
 OPEN_AI_KEY = None
 
-loader = RagGenerationLoader(
+loader = RagLoader(
     col_question= 'question', 
     col_context='context', 
     col_answer= 'answer', 
     col_label ='label')
-loader.load(INPUT_FILEPATH)
 
 # Faithfullness
-evaluator = RagFaithfulnessEvaluator(loader)
+loader.load(INPUT_FILEPATH)
+evaluator = FaithfulnessEvaluator(loader)
 evaluator.run()
 
 # Context Relevance
 INPUT_FILEPATH = 'data/rag/context_relevance_sample.json'
-
-loader = RagGenerationLoader(
+loader = RagLoader(
     col_question= 'question', 
     col_context='context', 
     col_answer= 'answer', 
     col_label ='label')
 loader.load(INPUT_FILEPATH)
-
-evaluator = RagContextRelevanceEvaluator(loader)
+evaluator = ContextRelevanceEvaluator(loader=loader)
 evaluator.run()
 
 # Answer Relevance
 
 INPUT_FILEPATH = 'data/rag/answer_relevance_sample.json'
-
-loader = RagGenerationLoader(
+loader = RagLoader(
     col_question= 'question', 
     col_context='context', 
     col_answer= 'answer', 
     col_label ='label')
 loader.load(INPUT_FILEPATH)
-
-evaluator = RagAnswerRelevanceEvaluator(loader)
+evaluator = AnswerRelevanceEvaluator(loader)
 evaluator.run()
-
